src/main/webapp/app/entities/service-data/route/service-data-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IServiceData, ServiceData } from '../service-data.model';
import { ServiceDataService } from '../service/service-data.service';

import { ServiceDataRoutingResolveService } from './service-data-routing-resolve.service';

describe('ServiceData routing resolve service', () => {
    let mockRouter: Router;
    let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
    let routingResolveService: ServiceDataRoutingResolveService;
    let service: ServiceDataService;
    let resultServiceData: IServiceData | undefined;

    beforeEach(() => {
        TestBed.configureTestingModule({
            imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
            providers: [
                {
                    provide: ActivatedRoute,
                    useValue: {
                        snapshot: {
                            paramMap: convertToParamMap({}),
                        },
                    },
                },
            ],
        });
        mockRouter = TestBed.inject(Router);
        jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
        mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
        routingResolveService = TestBed.inject(ServiceDataRoutingResolveService);
        service = TestBed.inject(ServiceDataService);
        resultServiceData = undefined;
    });

    describe('resolve', () => {
        it('should return IServiceData returned by find', () => {
            // GIVEN
            service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
            mockActivatedRouteSnapshot.params = { id: 123 };

            // WHEN
            routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
                resultServiceData = result;
            });

            // THEN
            expect(service.find).toBeCalledWith(123);
            expect(resultServiceData).toEqual({ id: 123 });
        });

        it('should return new IServiceData if id is not provided', () => {
            // GIVEN
            service.find = jest.fn();
            mockActivatedRouteSnapshot.params = {};

            // WHEN
            routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
                resultServiceData = result;
            });

            // THEN
            expect(service.find).not.toBeCalled();
            expect(resultServiceData).toEqual(new ServiceData());
        });

        it('should route to 404 page if data not found in server', () => {
            // GIVEN
            jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as ServiceData })));
            mockActivatedRouteSnapshot.params = { id: 123 };

            // WHEN
            routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
                resultServiceData = result;
            });

            // THEN
            expect(service.find).toBeCalledWith(123);
            expect(resultServiceData).toEqual(undefined);
            expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
        });
    });
});
